+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Not Released
 ---------------
 
+Breaking changes: index_of, last_index_of return nil instead of -1 if element can't be found
+
+Added:
+
+* find_index
 
 Release 0.1.0
 ---------------